+++ src/Bootstrap.php
@@ -9,10 +9,10 @@ use Hirtz\Location\Google\Components\Autocomplete;
 use Hirtz\Location\Google\Components\GoogleMapsApi;
 use Hirtz\Location\Models\Location;
 use Hirtz\Location\Modules\Admin\Widgets\Forms\LocationProviderIdField;
+use Hirtz\Skeleton\Helpers\EventHelper;
 use Hirtz\Skeleton\Web\Application;
 use Yii;
 use yii\base\BootstrapInterface;
-use yii\base\Event;
 use yii\db\BaseActiveRecord;
 
 class Bootstrap implements BootstrapInterface
@@ -24,7 +24,7 @@ class Bootstrap implements BootstrapInterface
     {
         Yii::setAlias('@location-google', __DIR__);
 
-        Event::on(Location::class, BaseActiveRecord::EVENT_INIT, $this->attachLocationProviderIdBehavior(...));
+        EventHelper::on(Location::class, BaseActiveRecord::EVENT_INIT, $this->attachLocationProviderIdBehavior(...));
 
         $googleApiKey = $app->params['googleApiKey'] ?? null;
 
@@ -36,10 +36,8 @@ class Bootstrap implements BootstrapInterface
         $this->setAutocompleteInputLabel();
     }
 
-    protected function attachLocationProviderIdBehavior(Event $event): void
+    protected function attachLocationProviderIdBehavior(Location $location): void
     {
-        /** @var Location $location */
-        $location = $event->sender;
         $location->attachBehavior('LocationProviderIdBehavior', LocationProviderIdBehavior::class);
     }